standardize `IsSortedKernel` and `MinMaxKernel` for the list encodings
Also see #4842.
