+++ vortex-array/src/arrays/list/compute/is_sorted.rs
@@ -3,21 +3,20 @@
 
 use vortex_error::VortexResult;
 
-use crate::arrays::{FixedSizeListArray, FixedSizeListVTable};
+use crate::arrays::{ListArray, ListVTable};
 use crate::compute::{IsSortedKernel, IsSortedKernelAdapter};
 use crate::register_kernel;
 
-/// IsSorted implementation for [`FixedSizeListArray`].
-impl IsSortedKernel for FixedSizeListVTable {
-    fn is_sorted(&self, _array: &FixedSizeListArray) -> VortexResult<Option<bool>> {
+impl IsSortedKernel for ListVTable {
+    fn is_sorted(&self, _array: &ListArray) -> VortexResult<Option<bool>> {
         // This would require comparing lists lexicographically.
         Ok(None)
     }
 
-    fn is_strict_sorted(&self, _array: &FixedSizeListArray) -> VortexResult<Option<bool>> {
+    fn is_strict_sorted(&self, _array: &ListArray) -> VortexResult<Option<bool>> {
         // This would require comparing lists lexicographically without duplicates.
         Ok(None)
     }
 }
 
-register_kernel!(IsSortedKernelAdapter(FixedSizeListVTable).lift());
+register_kernel!(IsSortedKernelAdapter(ListVTable).lift());
