+++ vortex-array/src/arrays/list/compute/min_max.rs
@@ -3,16 +3,15 @@
 
 use vortex_error::VortexResult;
 
-use crate::arrays::{FixedSizeListArray, FixedSizeListVTable};
+use crate::arrays::{ListArray, ListVTable};
 use crate::compute::{MinMaxKernel, MinMaxKernelAdapter, MinMaxResult};
 use crate::register_kernel;
 
-/// MinMax implementation for [`FixedSizeListArray`].
-impl MinMaxKernel for FixedSizeListVTable {
-    fn min_max(&self, _array: &FixedSizeListArray) -> VortexResult<Option<MinMaxResult>> {
+impl MinMaxKernel for ListVTable {
+    fn min_max(&self, _array: &ListArray) -> VortexResult<Option<MinMaxResult>> {
         // This would require finding the lexicographically minimum and maximum lists.
         Ok(None)
     }
 }
 
-register_kernel!(MinMaxKernelAdapter(FixedSizeListVTable).lift());
+register_kernel!(MinMaxKernelAdapter(ListVTable).lift());
